fix(pipex): patch bonus glitch
