+++ .vscode/settings.json
@@ -10,6 +10,7 @@
         "initializer_list": "c",
         "span": "c",
         "fdf.h": "c",
-        "libft.h": "c"
+        "libft.h": "c",
+        "pipex_bonus.h": "c"
     }
 }
