+++ milestone-2-pipex/pipex_bonus.h
@@ -1,0 +1,22 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   pipex_bonus.h                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: elagouch <elagouch@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/02/21 14:40:57 by elagouch          #+#    #+#             */
+/*   Updated: 2025/02/21 14:41:21 by elagouch         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef PIPEX_BONUS_H
+# define PIPEX_BONUS_H
+
+# ifndef PIPEX_PIPES
+#  define PIPEX_PIPES
+#  define PIPEX_PIPES_MIN 1
+#  define PIPEX_PIPES_MAX 1024
+# endif
+
+#endif
