+++ milestone-2-pipex/pipex_mandatory.h
@@ -1,0 +1,22 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   pipex_mandatory.h                                  :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: elagouch <elagouch@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/02/21 14:40:57 by elagouch          #+#    #+#             */
+/*   Updated: 2025/02/21 14:42:43 by elagouch         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef PIPEX_MANDATORY_H
+# define PIPEX_MANDATORY_H
+
+# ifndef PIPEX_PIPES
+#  define PIPEX_PIPES
+#  define PIPEX_PIPES_MIN 2
+#  define PIPEX_PIPES_MAX 2
+# endif
+
+#endif
